WIP prog: add a notion of a mutation "heatmap"
In this context, a heatmap is a flexible mechanism to choose where to
mutate some collection of bytes with some underlying probability
distribution, which could be dependent on the bytes themselves.

This commit defines the interface for heatmaps, and implements a generic
one.

The expected usage of the heatmap inferface is:
  - Heatmap type is selected based on program type (or other logic).
    Currently there is only one, but over time we will add more with differing
    probability distributions. To select a type, we simply create the
    appropriate empty struct.
  - Calculate the probability distribution by calling `hm.Populate(data)`.
  - Choose an index into the data wrt the heatmap distribution by
    calling `hm.Location(r)` for some random source `r`.

The generic heatmap is based on current compression for disk images. It
avoids mutating within long runs of constant bytes by viewing the file
as a series of 64-byte chunks, and ignoring chunks which consist of a
single repeated byte.