prog, pkg/compiler: add `BufferCompressed` buffer type & `compressed_image` builtin
Create the `BufferCompressed` kind of `BufferType`, which will be used
to represent compressed data. Create the corresponding `compressed_image`
syzlang builtin, which is backed by `BufferCompressed`. For now, no
syscalls use this feature - this will be introduced in future commits.

`BufferCompressed` arguments are mutated using a generic heatmap. In
future, we could add variants of `BufferCompressed` or populate the
`BufferType` sub-kind, using it to choose different kinds of heatmap for
different uncompressed data formats.

Various operations on compressed data must be forbidden, so the
`func (*BufferType) IsCompressed()` check avoids this in key places.
We could instead use equality checks on the `BufferKind`, but this
approach simplifies future modifications. Two key operations to avoid
are generation and minimisation - the compiler ensures that any syscalls
with `IsCompressed()` arguments must be marked `no_{generate,minimize}`.
When we do actually want to get/set the underlying uncompressed data,
the `DataArg` getter/setter transparently deal with the (de)compression.
We have to be careful *not* to decompress when serialising for execution
- this is the responsibility of the target pseudo-syscall.