executor: add `zlib` decompression header file
Create a header file to provide a clean entrypoint `decompress()`, which
decompresses data from an array to a file using `zlib`. This will be
used for pseudo-syscalls which accept compressed data, e.g.
`syz_mount_image`. The implementation is based on https://zlib.net/zpipe.c.

We have to ensure that `zlib` is correctly linked. Pass in the `-lz`
flag for this to be picked up. This commit also updates `Dockerfile`s to include
the `zlib1g` (`zlib` runtime) and `zlib1g-dev` (header files) packages.

TODO: provide `zlib` without relying on packages.