sys/linux: update asset storage for new `syz_mount_image`
Asset storage is now significantly simpler: we just take the
Base64-encoded, gzip-compressed image and output it (still compressed)
to a file. This commit removes most of the logic from `init_images.go`,
and therefore most of the tests from `init_images_test.go`.

We could instead keep this logic around and use it to adapt old-style
`syz_mount_image` calls in existing corpuses to match the new format.