sys/linux, tools/syz-imagegen: update `syz_mount_image` to use `compressed_image`
Rather than accepting "segments", `syz_mount_image` now accepts a
gzip-compressed image. Since this is already a corpus-breaking change,
also rearrange the arguments so that the image is at the end. This makes
it easier to inspect what the other arguments are set to.

This commit updates the descriptions and the `syz-imagegen` tool.
The executor, seed images, and asset saving will be updated in future commits.